requirements.txt
pandas
psycopg2-binary
sqlalchemy
tqdm
google-cloud-bigquery
clickhouse-driver[lz4]
google-cloud-storage
pyarrow 
phik
seaborn
polars
pyan3
scikit-learn
clickhouse-connect
pandas-gbq>=0.26.1
google-cloud-bigquery-storage
numpy
pysqlite3-binary